Render alm output format for Alignments using Mako
This is the first example of an output format created with a Mako
template. While whitespace handling for such formats can be a bit
tricky (note for example the \ after the python block in `alm_alm.mako`
to suppress the newline in the output), I still think the separation
of output creation and data becomes more transparent now.

Once we move to adapter-based output creation, this separation will
become even clearer, because then the two helper functions passed
to the template (`get_cogids` and `get_alm_string`) would be methods
of an AlignmentsAsAlm adapter, and the code in `Alignments.output` will
be reduced to adapter lookup and one render call.